+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Enhancements:**
 
+* Improved Email notifications UI, no more template, it is is now unified: has avatar, author name, reply comment data, etc
 * Added new option to hold only first comment and any further can pass without pre moderation, #353 
 * Added ability to import comments from HyperComments inside "Tools" tab by specifying URL to XML file, #326
 * Migrated Google+ to Google People API, #343